+++ src/renderer/src/env.d.ts
@@ -1,13 +1,8 @@
+import type { LanpmApi } from '../../shared/lanpm-api'
+
 declare global {
   interface Window {
-    lanpm: {
-      platform: NodeJS.Platform
-      versions: {
-        node: string
-        chrome: string
-        electron: string
-      }
-    }
+    lanpm: LanpmApi
   }
 }